Refactor providers into composables with global state
