--- src/providers/index.ts
@@ -1,10 +1,0 @@
-/* eslint-disable simple-import-sort/exports */
-/**
- * Order is important here!
- * If a provider is dependent on another it must be imported/exported below
- * the dependency.
- */
-export { default as UserSettingsProvider } from './user-settings.provider';
-export { default as TokenListProvider } from './token-lists.provider';
-export { default as TokensProvider } from './tokens.provider';
-export { default as AppProvider } from './app.provider';
