+++ src/providers/user-settings.provider.ts
@@ -1,17 +1,8 @@
 import { parseUnits } from '@ethersproject/units';
-import {
-  computed,
-  ComputedRef,
-  InjectionKey,
-  provide,
-  reactive,
-  Ref,
-  toRefs,
-} from 'vue';
+import { computed, reactive, toRefs } from 'vue';
 
 import { FiatCurrency } from '@/constants/currency';
 import LS_KEYS from '@/constants/local-storage.keys';
-import symbolKeys from '@/constants/symbol.keys';
 import { lsGet, lsSet } from '@/lib/utils';
 
 /**
@@ -22,20 +13,9 @@ export interface UserSettingsState {
   slippage: string;
 }
 
-export interface UserSettingsProviderResponse {
-  currency: Ref<FiatCurrency>;
-  slippage: Ref<string>;
-  slippageScaled: ComputedRef<string>;
-  setCurrency: (newCurrency: FiatCurrency) => void;
-  setSlippage: (newSlippage: string) => void;
-}
-
 /**
  * SETUP
  */
-export const UserSettingsProviderSymbol: InjectionKey<UserSettingsProviderResponse> =
-  Symbol(symbolKeys.Providers.App);
-
 const lsCurrency = lsGet(LS_KEYS.UserSettings.Currency, FiatCurrency.usd);
 const lsSlippage = lsGet(LS_KEYS.App.TradeSlippage, '0.01');
 
@@ -71,18 +51,12 @@ function setSlippage(newSlippage: string): void {
  * UserSettingsProvider
  * Provides global user settings interface
  */
-export default {
-  name: 'UserSettingsProvider',
-
-  setup(props, { slots }) {
-    provide(UserSettingsProviderSymbol, {
-      ...toRefs(state),
-      // methods
-      setCurrency,
-      setSlippage,
-      slippageScaled,
-    });
-
-    return () => slots.default();
-  },
-};
+export function getUserSettingsProvision() {
+  return {
+    ...toRefs(state),
+    // methods
+    setCurrency,
+    setSlippage,
+    slippageScaled,
+  };
+}
